test(snowflake): merge query-optimization tests into test_snowflake_queries.py
The optimization tests covered _build_pattern_filter / _make_composable — the
same units already tested by TestBuildPatternFilter/Schema/Table/View in
test_snowflake_queries.py. Splitting them by an arbitrary 'is it an
optimization' axis let the two files drift (a composition change updated one
file's expectations while silently breaking the other's).

Move all classes into the single home that already owns these query builders,
delete the separate file. No test logic changes — pure relocation.
